Move file transfer into kolibri.utils.
